Add esexamreport class.
This is a class to typeset external examiners' reports.  It produces
exceedingly ugly output, but that's what is imposed.  One possible
problem to deal with is that there is no way to override the year in the
heading.

(This is also my first attempt at writing a class file in the
documented-TeX literate programming style.  It seems to work.)