Include identifying information on subsequent pages of memos
We make the header include subject, date and page numbers, on the basis
that multiple-page memos are more likely to have pages go astray.
Possibly the first page should conditionally have page numering if it is
not also the last page, but that's a little ambitious for my current
state of TeX knowledge.