Better scoping for some commands in gsexam
Firstly, we can straightforwardly scope the enumeration labels and
references, by placing the definitions inside the question environment.
Since we have a minimal class file, in fact there are no outer
definitions, but that's fine for the straightforward use case of an exam
where all the material is either the frontmatter or within questions.

Secondly, arrange \clearpages at the beginning and end of the bscexam
environment, attempting to ensure that it's safe to use the metadata
commands (e.g. \coursecode) within the bscexam environment rather than
in the document preamble (i.e. at local rather than global scope).  The
motivation here is to make it possible to have multiple bscexam
environment (and so potentially multiple cover pages) within the same
document, for example in subject guides, or in a document collecting
multiple past papers.

Use this new style of metadata definition within the test document
provided.