Much ado about whitespace.
Don't start each question with a \clearpage any more.  The reasoning: we
will need to support multiple parts to an exam, and it might be silly to
have "PART A" all alone on a page.  (It might not, too; dunno).

Put a \clearpage in the definition of \coverpage, however, so that it
doesn't look silly.

The moment before we had that \clearpage in \coverpage revealed that I
don't really understand all the length specifications for lists.
Rearrange stuff so that "Question n" doesn't end up hanging into the
left margin.