Begin separation of gsexam into sty and cls
The idea here is that we may well want to have exams inside other
documents.  In such circumstances, the jobname-based answer handling,
headers and footers, and so on are inappropriate: these clearly belong
in a LaTeX class; however, the basic formatting of the environments is
generally applicable, and so belongs in a style file (which the class
will load using \RequirePackage).

This initial step separates out the handling of answers, noanswers and
the jobname-based answer handling.